pyproject.toml
[tool.poetry]
name = "affine"
version = "0.1.0"
description = "Thesaurus which accesses the Datamuse API."
authors = ["David Shivaji davidshivaji@mail.com"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
python-datamuse = "1.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
